requirements.txt
python==3.7.10
numpy==1.19.5
cv2==4.5.1
tensorflow==2.4.1
matplotlib==3.4.1
torch==1.7.0
os
shutil
time
RPi.GPIO